+++ src/main/java/org/opentripplanner/graph_builder/issues/RepeatedStops.java
@@ -1,0 +1,25 @@
+package org.opentripplanner.graph_builder.issues;
+
+import gnu.trove.list.TIntList;
+import org.opentripplanner.graph_builder.DataImportIssue;
+import org.opentripplanner.model.Trip;
+
+public class RepeatedStops implements DataImportIssue {
+
+    public static final String FMT = "Trip %s visits stops repeatedly. Removed duplicates at stop sequence numbers %s.";
+    
+    public final Trip trip;
+
+    public final TIntList removedStopSequences;
+    
+    public RepeatedStops(Trip trip, TIntList removedStopSequences){
+    	this.trip = trip;
+        this.removedStopSequences = removedStopSequences;
+    }
+    
+    @Override
+    public String getMessage() {
+        return String.format(FMT, trip.getId(), removedStopSequences);
+    }
+
+}