+++ src/main/java/org/opentripplanner/graph_builder/issues/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Graph builder data import issues represent errors or exceptional conditions encountered
+ * during the graph building process. They contain descriptive messages and potentially
+ * references to the objects in the graph that they annotate which facilitate visualization and
+ * cataloging/mapping of problems.
+ */
+package org.opentripplanner.graph_builder.issues;