+++ src/main/java/org/opentripplanner/graph_builder/module/TransitToTaggedStopsModule.java
@@ -3,14 +3,15 @@ package org.opentripplanner.graph_builder.module;
 import com.google.common.collect.Iterables;
 import org.locationtech.jts.geom.Envelope;
 import org.opentripplanner.common.geometry.SphericalDistanceLibrary;
+import org.opentripplanner.graph_builder.DataImportIssueStore;
 import org.opentripplanner.graph_builder.services.GraphBuilderModule;
 import org.opentripplanner.routing.edgetype.StreetTransitLink;
 import org.opentripplanner.routing.graph.Edge;
 import org.opentripplanner.routing.graph.Graph;
 import org.opentripplanner.routing.graph.Vertex;
 import org.opentripplanner.routing.impl.StreetVertexIndexServiceImpl;
-import org.opentripplanner.routing.vertextype.TransitStopVertex;
 import org.opentripplanner.routing.vertextype.TransitStopStreetVertex;
+import org.opentripplanner.routing.vertextype.TransitStopVertex;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -49,7 +50,11 @@ public class TransitToTaggedStopsModule implements GraphBuilderModule {
     }
 
     @Override
-    public void buildGraph(Graph graph, HashMap<Class<?>, Object> extra) {
+    public void buildGraph(
+            Graph graph,
+            HashMap<Class<?>, Object> extra,
+            DataImportIssueStore issueStore
+    ) {
         LOG.info("Linking transit stops to tagged bus stops...");
 
         index = new StreetVertexIndexServiceImpl(graph);
@@ -75,6 +80,8 @@ public class TransitToTaggedStopsModule implements GraphBuilderModule {
                 boolean wheelchairAccessible = ts.hasWheelchairEntrance();
                 if (!connectVertexToStop(ts, wheelchairAccessible)) {
                     LOG.debug("Could not connect " + ts.getStop().getCode() + " at " + ts.getCoordinate().toString());
+
+                    // TODO OTP2 - Why is this commented out? Is it not a problem or is it to nosey?
                     //LOG.warn(graph.addBuilderAnnotation(new StopUnlinked(ts)));
                 }
             }